# Add bitwise sign calculation for pauliprod improving performance