Added a bitwise implementation for pauli product sign computation, replacing the levi-cevita lookup table. This implementation removes the need for checking the changed indices and has most of its overhead from exponentiating the imaginary unit. It is implemented in such a way that the exponent itself can be calculated to save this overhead in hot loops where it may not be necessary to do the exponentiation each time.

The performance improvement can be seen from running the below code.
```
using BenchmarkTools, PauliPropagation

pstr1_dense = PauliString(128, rand([:I,:X,:Y,:Z], 128), 1:128)
pstr2_dense = PauliString(128, rand([:I,:X,:Y,:Z], 128), 1:128)

pstr1_sparse = PauliString(128, :X, 64)
pstr2_sparse = PauliString(128, :Z, 64)

@benchmark pauliprod($pstr1_dense, $pstr2_dense)
@benchmark pauliprod($pstr1_sparse, $pstr2_sparse)
```
The performance of my implementation is as follows (run on my laptop):

```
(dense)
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  23.383 ns … 128.074 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     23.412 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.009 ns ±   2.546 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▅      ▂▁                      ▂                            ▁
  ██▅▇▇▆▅▆██▆▅▅▅▃▄▆▅▅▅▆▇▅▅▅▅▄▅▆▅▅▅█▇▅▅▅▆▅▆▅▅▅▅█▆▃▂▃▄▄▄▃▄▄▃▂▄▃▄ █
  23.4 ns       Histogram: log(frequency) by time      32.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
```
(sparse)
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  15.804 ns … 35.685 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     16.430 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   16.741 ns ±  1.194 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

      █▅▆▇▄    ▁  ▁                        ▂  ▁▁         ▁    ▁
  ▇▅▆▂█████▆▅▄▅██▆██▇▇▇█▇▆▆▅▅▅▅▅▅▅▅▅▅▆▄▅▆▅▅██▇██▅▆▆▆▄▅▅▆▅█▆▆█ █
  15.8 ns      Histogram: log(frequency) by time      21.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

Old Version
```
(dense)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.278 μs …   2.985 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.300 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.332 μs ± 103.455 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄███▇▄     ▁ ▁                                 ▁ ▁▁▁     ▁  ▂
  ███████▆▅▄▇████▆▄▄▅▅▄▇▇█▇▆▅▆▅▅▅▄▅▅▅▅▄▃▃▅▅▄▅▄▄▄▇█████▇▆█████ █
  1.28 μs      Histogram: log(frequency) by time      1.71 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
```
(sparse no index)
BenchmarkTools.Trial: 10000 samples with 165 evaluations.
 Range (min … max):  645.000 ns …  1.111 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     656.742 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   671.659 ns ± 44.945 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▇██▇▅▄▂                                                     ▂
  ██████████▇█▇█▇▇▇▇▇▇▆▆▅▇▇▇▇▇█▇█▇▇▇████▇███████▇███▇█▇▆▅▅▅▆▄▅ █
  645 ns        Histogram: log(frequency) by time       852 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
```
(sparse with index)
@benchmark pauliprod($pstr1_sparse.term, $pstr2_sparse.term, 64)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  6.699 ns … 12.710 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.759 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.910 ns ±  0.508 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█▆▆▄▂    ▂▂                                   ▂▁     ▁▁   ▂
  ██████▇▇▄▆█████▅▃▃▄▄▅▆▅▅▅▃▅▄▄▄▁▁▄▅▅▃▃▅▅▇██▆▆▄▄███▇▆▇▆▅████ █
  6.7 ns       Histogram: log(frequency) by time     8.86 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

Note that this does degrade performance slightly for instances where the index is known, but it is small compared to the general case.  This is also a duplicate of #104, just submitted from a branch on this repository, rather than a fork.