@TysonRayJones 2025-06-18
Is it possible to explain the logic of `_calculatesignexponent()` similar to how functions like below are doc'd in `bitoperations.jl`?

https://github.com/MSRudolph/PauliPropagation.jl/blob/25f4c49538c70a1b61c4784f4087c75530830de2/src/PauliAlgebra/bitoperations.jl#L81-L96

Both for maintainability, _and_ because I'm curious as heck to know the secret sauce we'd missed!