@zmorrell 2025-06-18
Funny enough, the online Karnaugh map solver I used was not great, so I solved it by hand and found an additional mapping to an xor operation.  I think that the readability and efficiency is much better now.